Added some funcs

+++ dataservice/Models/Book.cs
@@ -10,7 +10,7 @@ namespace dataservice.Models
             Trainingsbook = new HashSet<Trainingsbook>();
         }
 
-        public int Isbn { get; set; }
+        public long Isbn { get; set; }
         public string Title { get; set; }
         public string Author { get; set; }
         public double Price { get; set; }
